Initial shot at an inline edit UI
Adds row actions in the ACM list table for View Conditionals, Edit, and Delete.

Connects each row's 'Edit' option to a hidden row containing a form with
all of the current Ad Code data.

Handles the basic Ad Code meta on a POST request, and updates accordingly,
but does not yet handle changes to the conditionals. Redirects back to the
referring site once saved.

Hopefully this approach isn't too far off, though I don't think it would
be too difficult to change course if it is.

View Conditionals and Delete are not yet hooked up to anything.

Side note - while preparing the conditionals, I noticed that multiple
arguments are possible for each. I tried saving an array() with multiple
arguments for a conditional, but it was still stored as a string. I didn't
investigate too much, but for now am displaying only argument 0 in the
inline display, as I didn't want to get distracted thinking out that
scenario.