Add Ability to delete conditionals via inline editor
Connect the hooks to delete conditionals through the inline editor. Allows
removal of all but one of the conditionals as the current requirement for
saving conditionals is that the incoming array not be empty. Whenever only
one conditional is present, the 'x' to delete does not show.

Changes some of the edit_conditionals code to not care about the existing
indexing as things are unpredictable otherwise. We could probably maintain
the index before form submission, but I don't think any problem arises
from just ignoring it.