Add filter to enable registration of external providers
This commit adds a filter - 'acm_register_provider_slug' - that enables a file included from a directory outside of the standard Ad Code Manager structure to be used as the provider extension class. We pass the existing provider data collected and allow another provider slug to be set with associated class information.

Closes #33