Add error handling for edit/ajax requests where a required parameter is missing
If we're performing an edit action via ajax, and one of the required parameters is missing, we can provide a WP Error object with a message that contains the missing parameter. We then return the error message to the front end via die(), which kills the ajax request as well and allows for another attempt.