Begin move of all Javascript into an anonymous function
Our Javascript is pretty haphazard right now with a few different approaches mixed in one. This starts the process of moving everything into one anonymous function. The script is now introduced in the footer as well so that we don't have to wait for document ready.

Breaking change due to `acm_add_more_conditionals` move, to be addressed in next commit.