Remove unused method post_admin_header
It looks like `post_admin_header()` was used to control ajax requests in the early days of Ad Code Manager, most likely before we switched to using list tables. The localized JS data here is unused anywhere else and hasn't been used by our Javascript in a while.

This commit removes that method as well as the action that called it.