Simplify defined constants
If the PHP magic `__FILE__` value for the plugin root file is captured, then all the values can be worked out as needed rather than abstracting into new constants.

The version constant isn't used by this plugin currently, but it may be in the future, and other code may want to know what version of the plugin is active.