Automate changelog from release notes
Props to @johnbillion for the work he had done on https://github.com/johnbillion/user-switching that this work is very minorly tweaked from.

- The README.md contains a HTML comment as a placeholder.
- During the deploy workflow (which happens when a non-pre-release release is made from a tag), it connects to the GitHub API, and pulls down previous release notes, removes any pre-release entries, sorts them into order, and loops through to create a string of headings and release notes.
- This then replaces the HTML comment in the README.md file.

Now instead of maintaining release notes in a CHANGELOG file or in the README, AND also in the GitHub release notes, it's now a single location (GitHub release notes).

The `changelog.js` workflow script can also be called locally, and this will do a one-time replacement of the HTML comment (for testing or removal of this automation). It needs a GitHub personal access token, and then call:

  TOKEN=... GITHUB_REPOSITORY='Automattic/ad-code-manager' node .github/workflows/changelog.js